refactor enhancedassumedstrains.hh
