+++ ikarus/finiteelements/mechanics/CMakeLists.txt
@@ -1,5 +1,6 @@
 # SPDX-FileCopyrightText: 2021-2024 The Ikarus Developers mueller@ibb.uni-stuttgart.de
 # SPDX-License-Identifier: LGPL-3.0-or-later
+add_subdirectory(eas)
 add_subdirectory(loads)
 add_subdirectory(materials)
 # install headers
@@ -11,5 +12,6 @@ install(
         kirchhoffloveshell.hh
         membranestrains.hh
         loads.hh
+        easvariants.hh
   DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/ikarus/finiteelements/mechanics
 )
